Make speed a read-only variable.
In order to make speed a read-only variable, we created a private field by prefixing the var name by underscore _, and provided a getter for it.

In Dart, this is not a breaking change, so we don't use setters and getters by default. We can start with a simple field, then update it to use custom setters and getters if you want to create read-only, write-only vars and compute or verify a value.